+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file, following t
 
 - PDBeMolstarPlugin.extensions.foldseek
 - Added options `leftPanel`, `rightPanel`, `logPanel`
+- Option `hideCanvasControls` accepts "trajectory" value (for multi-model structures)
 - All color options accept color names and hexcodes
+- `visualStyle` option allows per-component specification
+- StateGallery extension (work in progress)
 
 ## [v3.2.0] - 2024-04-24
 - Mol* core dependency updated to 3.45.0